+++ test/Elements.Tests/ProfileTests.cs
@@ -1,0 +1,51 @@
+using System;
+using Elements.Geometry;
+using Xunit;
+using System.Linq;
+using System.Collections.Generic;
+
+namespace Elements.Tests
+{
+    public class ProfileTests : ModelTest
+    {
+        [Fact]
+        public void Profile()
+        {
+            // Use the generated constructor to test that voids will
+            // be set to a default in the validator.
+            var profile = new Profile(Polygon.Rectangle(1, 1), null, default(Guid), null);
+            Assert.NotNull(profile.Voids);
+            profile.Voids.Add(Polygon.Rectangle(0.5, 0.5));
+        }
+
+        public void ProfileMultipleUnion()
+        {
+            this.Name = "MultipleProfileUnion";
+            // small grid of rough circles is unioned
+            // 2x3 grid shoud produce 2 openings
+            var circle = Polygon.Circle(3, 4);
+            var smallCircle = Polygon.Circle(1, 4);
+
+            var seed = new Profile(circle, new List<Polygon> { smallCircle }, Guid.NewGuid(), "");
+            for (int i = 0; i < 3; i++)
+            {
+                for (int j = 0; j < 2; j++)
+                {
+                    Transform t = new Transform(5 * i, 5 * j, 0);
+                    var newCircle = new Profile(new Polygon(circle.Vertices.Select(v => t.OfPoint(v)).ToArray()),
+                                                new List<Polygon> { new Polygon(smallCircle.Vertices.Select(v => t.OfPoint(v)).ToArray()) },
+                                                Guid.NewGuid(),
+                                                "");
+
+                    seed = seed.Union(newCircle);
+
+                    if (seed == null) throw new Exception("Making union failed");
+                }
+            }
+            var floor = new Floor(seed, 1);
+            this.Model.AddElement(floor);
+
+            Assert.Equal(8, seed.Voids.Count());
+        }
+    }
+}